+++ examples/offload_functions.js
@@ -1,11 +1,14 @@
 var workerpool = require('./../index');
 
+// create a worker pool
 var pool = workerpool.pool();
 
+// create a static function
 function add(a, b) {
   return a + b;
 }
 
+// offload execution of a function to the worker pool
 pool.run(add, [3, 4])
     .then(function (result) {
       console.log('result', result); // outputs 7