+++ HISTORY.md
@@ -2,6 +2,14 @@
 https://github.com/josdejong/workerpool
 
 
+## not yet released, version 1.0.0
+
+- Implemented support for both node.js and the browser.
+- Implemented offloading functions.
+- Implemented worker proxy.
+- Added docs and examples.
+
+
 ## 2014-05-02, version 0.0.1
 
 - Module name registered at npm.